chore: fixes build errors
